@@ -25,25 +25,33 @@
   },
   "homepage": "https://github.com/bingnz/gulp-less-changed#readme",
   "devDependencies": {
+    "@types/bluebird-global": "^3.0.1",
+    "@types/core-js": "^0.9.35",
+    "@types/gulp-util": "^3.0.30",
+    "@types/less": "https://registry.npmjs.org/@types/less/-/less-0.0.32.tgz",
+    "@types/mkdirp": "^0.3.29",
+    "@types/through2": "^2.0.32",
+    "@types/vinyl": "^2.0.0",
     "chai": "^3.5.0",
-    "coveralls": "^2.11.9",
+    "coveralls": "^2.11.15",
     "fake-fs": "^0.5.0",
-    "gulp": "^3.9.0",
-    "gulp-istanbul": "^0.10.3",
-    "gulp-load-plugins": "^1.2.0",
-    "gulp-mocha": "^2.2.0",
-    "gulp-sourcemaps": "^1.6.0",
-    "gulp-typescript": "^2.12.1",
-    "merge2": "^1.0.2",
-    "mocha": "^2.4.5",
+    "gulp": "^3.9.1",
+    "gulp-istanbul": "^1.1.1",
+    "gulp-load-plugins": "^1.4.0",
+    "gulp-mocha": "^3.0.1",
+    "gulp-replace": "^0.5.4",
+    "gulp-sourcemaps": "^2.2.1",
+    "gulp-typescript": "^3.1.3",
+    "lazypipe": "^1.0.1",
+    "merge2": "^1.0.3",
+    "mocha": "^3.2.0",
     "pmock": "^0.2.3",
-    "proxyquire": "^1.7.4",
-    "semantic-release": "^4.3.5",
-    "sinon": "^1.17.3",
+    "proxyquire": "^1.7.10",
+    "semantic-release": "^6.3.5",
+    "sinon": "^1.17.7",
     "sinon-chai": "^2.8.0",
     "stream-assert": "^2.0.3",
-    "through2": "^2.0.0",
-    "typescript": "^1.7.5"
+    "typescript": "^2.1.4"
   },
   "dependencies": {
     "bluebird": "^3.4.0",
