@@ -4,15 +4,16 @@ var merge = require('merge2');
 var plugins = require('gulp-load-plugins')();
 var through2 = require('through2');
 var os = require('os');
+var lazypipe = require('lazypipe');
 
 var tsProject = plugins.typescript.createProject('tsconfig.json', {
     typescript: require('typescript')
 });
 
 gulp.task('compile', function () {
-    var tsResult = gulp.src(['src/**/*.ts', 'src/*.d.ts', 'typings/index.d.ts'])
+    var tsResult = gulp.src(['src/**/*.ts', 'src/*.d.ts'])
         .pipe(plugins.sourcemaps.init())
-        .pipe(plugins.typescript(tsProject));
+        .pipe(tsProject());
 
     return merge([
         tsResult.dts
@@ -28,28 +29,16 @@ gulp.task('compile', function () {
     ]);
 });
 
-// http://stackoverflow.com/questions/22155106/typescript-code-coverage-of-multiple-extends-declarations/26321994#26321994
-function istanbulIgnoreTypeScriptExtend() {
-    var tsExtends = /var __extends =/;
-    return through2.obj(function(file, enc, done) {
-        if (file.isBuffer() && tsExtends.test(file.contents)) {
-            var rows = file.contents.toString().split('\n');
-            for (var i = 0; i < rows.length; i++) {
-                if (rows[i].indexOf('var __extends =') === 0) {
-                    rows.splice(i, 0, '/* istanbul ignore next: TypeScript extend */');
-                    break;
-                }
-            }
-            file.contents = new Buffer(rows.join(os.EOL));
-        }
-        this.push(file);
-        done();
-    });
+function istanbulTypeScriptIgnores() {
+    var ignoreReplacement = '/* istanbul ignore next */\n$1'
+    return lazypipe()
+        .pipe(plugins.replace, /(var __extends =)/, ignoreReplacement)
+        .pipe(plugins.replace, /(var _this = _super\.call)/, ignoreReplacement)();
 }
 
 gulp.task('pre-test', ['compile'], function () {
   return gulp.src(['release/**/*.js'])
-    .pipe(istanbulIgnoreTypeScriptExtend())
+    .pipe(istanbulTypeScriptIgnores())
     // Covering files
     .pipe(plugins.istanbul())
     // Force `require` to return covered files
